Prevent a NullPointerException of the parser on a channel poll.
When encountering a channel poll the parser calls recv_args with a null
parameter, probably because a channel poll is not supported anyway. This
causes a NullPointerException before the more descriptive
MyParseException can be thrown. The MyParseException can actually be
thrown immediately after parsing the opening bracket, similar to the
other unsupported language features.