Invalid declarations of mtype should still be handled by the mtype grammar.
The previous use of LOOKAHEAD caused erroneous declarations of mtype to
be parsed by the one_decl grammar. (Resulting in unintuitive error
messages.) Giving priority to the mtype grammar by listing it before
one_decl already resolves the problem. JavaCC selects from a list of
parsing options by going from top to bottom and choosing the first
match. Cases of ambiguity are therefore handled automatically. Still,
the JavaCC compiler will warn the user. By using LOOKAHEAD(1) we inform
the parser that we are aware of the situation.