Add support for arrays of channels.
Promela variables of the model are mapped to Java variables in the
verifier. Channels are handled similar to variables, but instead of the
channel itself, the variable stores an index to be used in the _channel
array of the verifier.

So individual channels are referenced using Variables, but this does not
take into account that a Variable can store an array of values. What
channel is being referenced when we have an array of indices for the
_channel array?

Instead of a Variable an Identifier should be used, which is a reference
to a specific element of a Variable.

Exclusive send and receive (XS and XR) are still defined over the
Variable as a whole, as per the language definition.