Ensure the never claim model detects else and end transitions.
Never claim actions are interleaved with model actions using transition
take events. These events should be signaled on the model as a whole,
and not a specific process, since listeners only register with the
model. The never claim and model will otherwise not be interleaved
correctly, possibly preventing detecting of never claim violations.