Ensure that processes get a correct pid when decoding states.
When decoding a PromelaModel state, processes of the old state will be
reused whenever possible (instead of constructing a new PromelaProcess
object). This requires the process in the new state to at least have the
same proctype and pid as in the old one.

The proctype of the new and old process at a given pid can differ. In
that case a new PromelaProcess is created, using the same methods as
during normal initialisation and execution (i.e. when taking a run
transition).

However, when decoding, the _nrProcs and _nrChannels variables in the
PromelaModel do not have the correct values for constructing a new
proccess. This can cause new processes to get wrong _pid and
_nrChannelBefore values.

Instead of implicitly obtaining these values from the model state, they
should be passed explicitly when creating a new process (_pid), or be
obtained through other means (_nrChannelsBefore).