Parameters should always be encoded, even when not written to.
Variables are not encoded in the state vector when they are never
written to. Standard initialisation of the variable is assumed to assign
the right value.

In the case of parameters however, the value is assigned at runtime when
instantiating the proctype. Even if parameters are never explicitely
modified (by assigning a value), they should still be encoded.