Fixed the partial implementation of calculating depth for BFS.
The Queue used by a BFS already has a partial implementation for
recording the search depth. A special nextDepth state, is used as a barrier
for calculating the depth reached.

However, this state was being returned to the search algorithm as
a genuine state. This caused an incorrect numbers transitions being
explored by the BFS.

The generic SearchAlgorithm class also calculated the depth in a manner
specific to DFS, covering up a small error in DepthFirstSearch.