Detect invalid end states and correctly calculate their depth in a BFS.
Because of an unintentional swap of two parameters, a BFS would never
detect invalid end states in a default search.

Also, the depth reached in a BFS is not the same as the depth for which
states have been generated as in a DFS. This in turn causes problems
when an invalid end state is detected before updating the maximum depth.